@@ -20,23 +20,23 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "5.2.3",
-    "@angular/cdk": "^5.0.4",
-    "@angular/common": "5.2.3",
-    "@angular/compiler": "5.2.3",
-    "@angular/core": "5.2.3",
-    "@angular/flex-layout": "^2.0.0-beta.12",
-    "@angular/forms": "5.2.3",
-    "@angular/http": "5.2.3",
-    "@angular/material": "^5.0.4",
-    "@angular/material-moment-adapter": "^5.0.4",
-    "@angular/platform-browser": "5.2.3",
-    "@angular/platform-browser-dynamic": "5.2.3",
-    "@angular/router": "5.2.3",
+    "@angular/animations": "6.0.2",
+    "@angular/cdk": "6.0.2",
+    "@angular/common": "6.0.2",
+    "@angular/compiler": "6.0.2",
+    "@angular/core": "6.0.2",
+    "@angular/flex-layout": "^6.0.0-beta.15",
+    "@angular/forms": "6.0.2",
+    "@angular/http": "6.0.2",
+    "@angular/material": "6.0.2",
+    "@angular/material-moment-adapter": "^6.0.2",
+    "@angular/platform-browser": "6.0.2",
+    "@angular/platform-browser-dynamic": "6.0.2",
+    "@angular/router": "6.0.2",
     "@angular-devkit/core": "^0.0.29",
     "@ngx-translate/core": "9.0.2",
     "@ngx-translate/http-loader": "^2.0",
-    "angular2-notifications": "^0.9.7",
+    "angular2-notifications": "^1.0.2",
     "basil.js": "^0.4.5",
     "core-js": "^2.4.1",
     "hammerjs": "2.0.8",
@@ -44,21 +44,22 @@
     "jwt-simple": "^0.5.1",
     "material-design-icons": "^3.0.1",
     "moment": "^2.17.1",
-    "ngx-md": "3.1.0",
+    "ngx-md": "3.1.1",
     "object-assign": "^4.1.1",
     "raven-js": "3.21.0",
-    "rxjs": "5.5.6",
+    "rxjs": "6.1.0",
+    "rxjs-compat": "6.1.0",
     "typeface-montserrat": "^0.0.45",
     "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.4"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angular/cli": "^1.6",
-    "@angular/compiler-cli": "5.2.3",
+    "@angular/compiler-cli": "6.0.2",
     "@types/jasmine": "2.8.3",
     "@types/jwt-simple": "^0.5.33",
     "@types/node": "~9.3.0",
-    "codelyzer": "^4.1",
+    "codelyzer": "^4.3.0",
     "jasmine-core": "^2.8.0",
     "jasmine-spec-reporter": "^4.2.1",
     "karma": "^2.0.0",
@@ -73,6 +74,6 @@
     "protractor": "~5.2.2",
     "ts-node": "^4.1.0",
     "tslint": "^5.9.1",
-    "typescript": "^2.6.0"
+    "typescript": "2.7.2"
   }
 }
