@@ -1,6 +1,6 @@
 import {TranslateLoader} from '@ngx-translate/core';
 import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 
 @Injectable()
 export class MockTranslateLoader implements TranslateLoader {
