@@ -1,5 +1,5 @@
 import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {ColorService} from '../services/color.service';
 import {ColorDto} from '../model/color.dto';
 
