@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {MockCrudService} from '../core/mock-crud.service';
 import {TaskDto} from '../../../task/model/task.dto';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {SchoolClassSubjectDto} from '../../../task/model/school-class-subject.dto';
 import {MockSubjectService} from '../manage/manage-entities/mock-subject.service';
 import * as moment from 'moment';
