@@ -1,6 +1,6 @@
 import {MediaChange, ObservableMedia} from '@angular/flex-layout';
 import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {Subscription} from 'rxjs/Subscription';
 
 @Injectable()
