@@ -3,14 +3,15 @@ import {Router} from '@angular/router';
 import {AuthService} from '../../interfaces/auth.service';
 import * as Raven from 'raven-js';
 import {IdentityProvider} from './identity-provider';
-import {Observable} from 'rxjs/Observable';
+import {Observable, of, throwError} from 'rxjs';
 import {UsernamePasswordDto} from '../../../auth/model/username-password.dto';
 import {environment} from '../../../../environments/environment';
 import {JwtHelperService} from './jwt-helper.service';
 import {AuthResponseDto} from './auth-response.dto';
 import {BasilWrapperService} from '../../persistence/basil-wrapper.service';
 import {HttpClient, HttpHeaders} from '@angular/common/http';
-import {IdTokenDto} from "./id-token.dto";
+import {IdTokenDto} from './id-token.dto';
+import {map} from 'rxjs/operators';
 
 @Injectable()
 export class DefaultAuthService implements AuthService {
@@ -63,22 +64,22 @@ export class DefaultAuthService implements AuthService {
     private _handleUsernamePasswordAuth(usernamePassword: UsernamePasswordDto, isSignUp: boolean): Observable<boolean> {
         this._removeTokenIfExpired();
         if (this.isLoggedIn()) {
-            return Observable.of(true);
+            return of(true);
         }
         return this._http.post<AuthResponseDto>(`/api/auth/${isSignUp ? 'signUp' : 'login'}`, usernamePassword)
-            .map(token => this._afterLogin(token));
+            .pipe(map(token => this._afterLogin(token)));
     }
 
     private _handleIdentityProviderAuth(identityProvider: IdentityProvider, isSignUp: boolean, token: string): Observable<boolean> {
         this._removeTokenIfExpired();
         if (this.isLoggedIn()) {
-            return Observable.of(true);
+            return of(true);
         }
         if (identityProvider == IdentityProvider.GOOGLE) {
             return this._http.post<AuthResponseDto>(`/api/auth/${isSignUp ? 'signUp' : 'login'}/google/`, <IdTokenDto>{idToken: token}, {headers: new HttpHeaders({'Content-Type': 'application/json'})})
-                .map(token => this._afterLogin(token));
+                .pipe(map(token => this._afterLogin(token)));
         }
-        return Observable.throw(new Error('Identity provider not supported'));
+        return throwError(new Error('Identity provider not supported'));
     }
 
     private _removeTokenIfExpired() {
