@@ -1,8 +1,10 @@
 import {Cacheable} from '../../interfaces/cacheable';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {Util} from '../../util/util';
 import {AppService} from './app.service';
 import {HttpClient} from '@angular/common/http';
+import {of} from 'rxjs/internal/observable/of';
+import {map, share} from 'rxjs/operators';
 
 /**
  * Default implementation of a caching service
@@ -23,19 +25,21 @@ export abstract class CacheableService<T> extends AppService implements Cacheabl
      * @param baseUri of the services http calls
      * @param expiredAfter time in milliseconds that the cache lives (default 10 minutes)
      */
-    constructor(http: HttpClient = null, baseUri: string = '', private expiredAfter: number = 600000) {
+    protected constructor(http: HttpClient = null, baseUri: string = '', private expiredAfter: number = 600000) {
         super(http, baseUri);
     }
 
     getFromCacheOrFetch(fetchFunc: () => Observable<T>) {
-        if (this.hasCache()) return Observable.of(this.cache);
+        if (this.hasCache()) return of(this.cache);
         else if (this.observable) return this.observable;
-        return this.saveObservable(fetchFunc()
-            .map((res: T) => {
+        return this.saveObservable(fetchFunc().pipe(
+            map((res: T) => {
                 this.clearObservable();
                 this.saveCache(res);
                 return this.cache;
-            }).share());
+            }),
+            share()
+        ));
     }
 
     /**
