@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 import {AppService} from './core/app.service';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {Info} from '../model/info.dto';
 import {HttpClient} from '@angular/common/http';
 
