@@ -7,7 +7,8 @@ import {isNull} from 'util';
 import {MomentDateAdapter} from '@angular/material-moment-adapter';
 import {UserService} from './user.service';
 import {isTrue} from '../util/helper';
-import {Observable} from 'rxjs/Observable';
+import {Observable, timer} from 'rxjs';
+import {map, switchMap} from 'rxjs/operators';
 
 export class OCValidators {
     public static readonly PASSWORD_REGEX: RegExp = /^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=])(?=\S+$).{8,}$/;
@@ -129,11 +130,12 @@ export class OCValidators {
 
     public static checkMailNotTaken(userService: UserService): AsyncValidatorFn {
         return (control: AbstractControl): Observable<ValidationErrors | null> => {
-            return Observable.timer(500)
-                .switchMap(() => userService.checkMailNotTaken(control.value))
-                .map(res => {
+            return timer(500).pipe(
+                switchMap(() => userService.checkMailNotTaken(control.value)),
+                map(res => {
                     return isTrue(res) ? null : {mailTaken: true};
-                });
+                })
+            );
         }
     }
 }
