@@ -1,7 +1,7 @@
 import {AppCrudService} from '../../core/services/core/app-crud.service';
 import {ExamTaskDto} from '../model/exam.task.dto';
 import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {HttpClient} from '@angular/common/http';
 
 /**
