@@ -6,7 +6,7 @@ import {OCUiModule} from '../../oc-ui/oc-ui.module';
 import {ReactiveFormsModule} from '@angular/forms';
 import {RouterTestingModule} from '@angular/router/testing';
 import {ActivatedRoute, convertToParamMap} from '@angular/router';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {MockMarkService} from '../../core/mock/mark/mock-mark.service';
 import {MarkValueComponent} from '../mark-value/mark-value.component';
 import {MarkHighlighterDirective} from '../mark-highlighter/mark-highlighter.directive';
