@@ -4,7 +4,7 @@ import {MarkSemesterComponent} from './mark-semester.component';
 import {TestModule} from '../../core/mock/test.module';
 import {OCUiModule} from '../../oc-ui/oc-ui.module';
 import {ActivatedRoute, convertToParamMap} from '@angular/router';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {MockMarkService} from '../../core/mock/mark/mock-mark.service';
 import {MarkValueComponent} from '../mark-value/mark-value.component';
 import {ReactiveFormsModule} from '@angular/forms';
