@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core';
 import {MarkDto} from '../model/mark.dto';
 import {SemesterMarkDto} from '../model/semester-mark.dto';
-import {Observable} from 'rxjs/Observable';
+import {Observable} from 'rxjs';
 import {AppService} from '../../core/services/core/app.service';
 import {MarkGroupDto} from '../model/mark-group.dto';
 import {HttpClient} from '@angular/common/http';
