@@ -1,8 +1,9 @@
 import {ActivatedRouteSnapshot, Resolve, RouterStateSnapshot} from '@angular/router';
-import {Observable} from 'rxjs/Observable';
+import {Observable, of} from 'rxjs';
 import {Injectable} from '@angular/core';
 import {MarkService} from './mark.service';
 import {MarkGroupDto} from '../model/mark-group.dto';
+import {map, switchMap} from 'rxjs/operators';
 
 @Injectable()
 export class SubjectMarkGroupResolver implements Resolve<MarkGroupDto> {
@@ -10,9 +11,10 @@ export class SubjectMarkGroupResolver implements Resolve<MarkGroupDto> {
     }
 
     resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): MarkGroupDto | Observable<MarkGroupDto> | Promise<MarkGroupDto> {
-        return Observable.of(route.paramMap)
-            .map(map => parseInt(map.get('subjectId')))
-            .switchMap(subjectId => this._markService.getMarkGroupBySubjectId(subjectId));
+        return of(route.paramMap).pipe(
+            map(map => parseInt(map.get('subjectId'))),
+            switchMap(subjectId => this._markService.getMarkGroupBySubjectId(subjectId))
+        )
     }
 
 }
