@@ -1,8 +1,9 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {ExamDto} from '../model/exam.dto';
 import {ExamTaskService} from '../service/exam-task.service';
-import {Subject} from 'rxjs/Subject';
+import {Subject} from 'rxjs';
 import {ExamTaskDto} from '../model/exam.task.dto';
+import {switchMap} from 'rxjs/operators';
 
 @Component({
     selector: 'exam-list-item',
@@ -21,8 +22,7 @@ export class ExamListItemComponent implements OnInit {
     }
 
     ngOnInit() {
-        this.finishTask$
-            .switchMap(id => this._examTaskService.changeState(id))
+        this.finishTask$.pipe(switchMap(id => this._examTaskService.changeState(id)))
             .subscribe(examTask => this._updateExamTaskList(examTask));
     }
 
