@@ -1,8 +1,8 @@
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
-import {InstitutionDto} from '../model/manage.dto';
-import {AppCrudService} from '../../../core/services/core/app-crud.service';
 import {HttpClient} from '@angular/common/http';
+import {AppCrudService} from '../../core/services/core/app-crud.service';
+import {InstitutionDto} from '../old/model/manage.dto';
 
 @Injectable()
 export class InstitutionService extends AppCrudService<InstitutionDto> {
