@@ -1,8 +1,8 @@
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
-import {SchoolClassDto} from '../model/manage.dto';
-import {AppCrudService} from '../../../core/services/core/app-crud.service';
 import {HttpClient} from '@angular/common/http';
+import {AppCrudService} from '../../core/services/core/app-crud.service';
+import {SchoolClassDto} from '../old/model/manage.dto';
 
 @Injectable()
 export class SchoolClassService extends AppCrudService<SchoolClassDto> {
