@@ -1,10 +1,10 @@
 import {Component, ElementRef, EventEmitter, Input, OnDestroy, OnInit, Output, ViewChild} from '@angular/core';
 import {TranslateService} from '@ngx-translate/core';
-import {Subject} from 'rxjs/Subject';
+import {fromEvent, Subject} from 'rxjs';
 import {FormBuilder, FormGroup} from '@angular/forms';
-import {Observable} from 'rxjs/Observable';
 import {ConnectionPositionPair} from '@angular/cdk/overlay';
 import {isEmpty, isNotEmpty} from '../../../core/util/helper';
+import {debounceTime, distinctUntilChanged, takeUntil} from 'rxjs/operators';
 
 @Component({
     selector: 'oc-filter-search',
@@ -51,22 +51,22 @@ export class OCFilterSearchComponent implements OnInit, OnDestroy {
         this._currentPlaceholder = this._getGivenOrPlaceholder(this.placeholder);
         this._refreshTriggerDimensions();
 
-        this.search$
-            .takeUntil(this._ngUnsubscribe)
-            .debounceTime(300)
-            .distinctUntilChanged()
-            .subscribe(searchStr => {
-                this.onSearch.emit(searchStr);
-                this._currentPlaceholder = this._getGivenOrPlaceholder(isNotEmpty(searchStr) ? searchStr : this.placeholder);
-            });
+        this.search$.pipe(
+            takeUntil(this._ngUnsubscribe),
+            debounceTime(300),
+            distinctUntilChanged()
+        ).subscribe(searchStr => {
+            this.onSearch.emit(searchStr);
+            this._currentPlaceholder = this._getGivenOrPlaceholder(isNotEmpty(searchStr) ? searchStr : this.placeholder);
+        });
 
-        Observable.fromEvent(window, 'resize')
-            .takeUntil(this._ngUnsubscribe)
-            .subscribe(() => this._refreshTriggerDimensions());
+        fromEvent(window, 'resize').pipe(
+            takeUntil(this._ngUnsubscribe)
+        ).subscribe(() => this._refreshTriggerDimensions());
 
-        this._translateService.onLangChange
-            .takeUntil(this._ngUnsubscribe)
-            .subscribe(() => {
+        this._translateService.onLangChange.pipe(
+            takeUntil(this._ngUnsubscribe)
+        ).subscribe(() => {
             if (isEmpty(this.placeholder) && isEmpty(this.searchForm.get('search').value)) {
                 this._currentPlaceholder = this._getDefaultPlaceholder().call(this);
             }
