@@ -1,7 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {ActiveSemesterService} from '../../../core/services/semester/active-semester.service';
 import {SemesterDto} from '../../../manage/old/model/manage.dto';
-import {Subject} from 'rxjs/Subject';
+import {Subject} from 'rxjs';
 
 @Component({
     selector: 'semester-chooser',
